flood-image-segmentation-ml4im-2024-2025/case-cline/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
scikit-learn>=1.0.0
tqdm>=4.60.0
opencv-python>=4.5.0
